+++ Project.toml
@@ -22,4 +22,7 @@ RecipesBase = "1"
 Unitful = "1"
 UnitfulAstro = "1"
 WCS = "0.6"
-julia = "1.6"
+julia = "1.10"
+
+[workspace]
+projects = ["test", "docs"]
