+++ docs/Project.toml
@@ -7,9 +7,6 @@ Spectra = "391af1a9-06f1-59d3-8d21-0be089654739"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 UnitfulAstro = "6112ee07-acf9-5e0f-b108-d242c714bf9f"
 
-[sources]
-Spectra = {path = ".."}
-
 [compat]
 Documenter = "1"
 UnitfulAstro = ">=0.4.0"
