+++ test/aqua.jl
@@ -1,0 +1,3 @@
+using Aqua: test_all
+
+test_all(Spectra)
